Added symbols for location drawing. Incl mono font. State Person info
Symbols can be attached to GameObjects to allow them to be draw on the map. Index gametick uses the enhanced location drawing. FreeMono Font was added due to need for a monospaced font that supports unicode characters.

Fix: Pace no longer gets out of sync when blocked.
Touchup: Information on the gamestate's People is displayed on page.